{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.27.0"
alembic = "^1.10.4"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.11"}
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.6"
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
sentry-sdk = "^1.35.0"
amplitude-analytics = "^1.1.3"

[tool.poetry.dev-dependencies]
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
